+++ sql-to-dbsp-compiler/SQL-compiler/src/test/java/org/dbsp/sqlCompiler/compiler/sql/postgres/PostgresInt4Tests.java
@@ -343,14 +343,14 @@ public class PostgresInt4Tests extends SqlIoTest {
         this.q(
                 """
                         SELECT x, x::int4 AS int4_value
-                        FROM (VALUES (-2.9::numeric),
-                                     (-2.5::numeric),
-                                     (-1.5::numeric),
-                                     (-0.5::numeric),
-                                     (0.0::numeric),
-                                     (0.5::numeric),
-                                     (1.5::numeric),
-                                     (2.5::numeric)) t(x);
+                        FROM (VALUES (-2.9::numeric(2, 1)),
+                                     (-2.5::numeric(2, 1)),
+                                     (-1.5::numeric(2, 1)),
+                                     (-0.5::numeric(2, 1)),
+                                     (0.0::numeric(2, 1)),
+                                     (0.5::numeric(2, 1)),
+                                     (1.5::numeric(2, 1)),
+                                     (2.5::numeric(2, 1))) t(x);
                           x   | int4_value
                         ------+------------
                          -2.9 |         -2